+++ boards/crkbd/kb.py
@@ -16,7 +16,7 @@ class KMKKeyboard(_KMKKeyboard):
     )
     row_pins = (board.P0_22, board.P0_24, board.P1_00, board.P0_11)
     diode_orientation = DiodeOrientation.COLUMNS
-    uart_pin = board.P0_08
+    data_pin = board.P0_08
     rgb_pixel_pin = board.P0_06
     i2c = board.I2C
     powersave_pin = board.P0_13
